t/xml/entities-2.dat

#data
<!DOCTYPE a[
<!ENTITY%>
]><a/>
#errors
2;9;no space before md name;;;m
2;10;no md name;;;m
#document
| <!DOCTYPE a>
| <{} a>
#entities

#data
<!DOCTYPE a[
<!ENTITY%
#errors
2;9;no space before md name;;;m
2;9;unclosed md;;;m
2;9;unclosed internal subset;;;m
2;9;no root element;;;m
#document
| <!DOCTYPE a>
#entities

#data
<!DOCTYPE a[
<!ENTITY %>
]><a/>
#errors
2;11;no md name;;;m
#document
| <!DOCTYPE a>
| <{} a>
#entities

#data
<!DOCTYPE a[
<!ENTITY %
#errors
2;10;unclosed md;;;m
2;10;unclosed internal subset;;;m
2;10;no root element;;;m
#document
| <!DOCTYPE a>
#entities

#data
<!DOCTYPE a[
<!ENTITY % bc PUBLIC "x" "y">
]><a/>
#errors
#document
| <!DOCTYPE a>
| <{} a>
#entities
| <!ENTITY % bc "" "x" "y" >

#data
<!DOCTYPE a[
<!ENTITY % bc PUBLIC "x""y">
]><a/>
#errors
2;25;no space before system literal;;;m
#document
| <!DOCTYPE a>
| <{} a>
#entities
| <!ENTITY % bc "" "x" "y" >

#data
<!DOCTYPE a[
<!ENTITY % bc PUBLIC "x"'y'>
]><a/>
#errors
2;25;no space before system literal;;;m
#document
| <!DOCTYPE a>
| <{} a>
#entities
| <!ENTITY % bc "" "x" "y" >

#data
<!DOCTYPE a[
<!ENTITY % bc PUBLIC 'x'"y">
]><a/>
#errors
2;25;no space before system literal;;;m
#document
| <!DOCTYPE a>
| <{} a>
#entities
| <!ENTITY % bc "" "x" "y" >

#data
<!DOCTYPE a[
<!ENTITY % bc PUBLIC 'x''y'>
]><a/>
#errors
2;25;no space before system literal;;;m
#document
| <!DOCTYPE a>
| <{} a>
#entities
| <!ENTITY % bc "" "x" "y" >

#data
<!DOCTYPE a[
<!ENTITY % bc PUBLIC "x" "y" NDATA z>
]><a/>
#errors
2;30;string after SYSTEM literal;;;m
#document
| <!DOCTYPE a>
| <{} a>
#entities
| <!ENTITY % bc "" "x" "y" >

#data
<!DOCTYPE a[
<!ENTITY % b "">
]><a/>
#errors
#document
| <!DOCTYPE a>
| <{} a>
#entities
<!ENTITY % b "" "" "" >

#data
<!DOCTYPE a[
<!ENTITY % b ''>
]><a/>
#errors
#document
| <!DOCTYPE a>
| <{} a>
#entities
<!ENTITY % b "" "" "" >

#data
<!DOCTYPE a[
<!ENTITY % b "xxxxx">
]><a/>
#errors
#document
| <!DOCTYPE a>
| <{} a>
#entities
<!ENTITY % b "xxxxx" "" "" >

#data
<!DOCTYPE a[
<!ENTITY % b 'xxxxx'>
]><a/>
#errors
#document
| <!DOCTYPE a>
| <{} a>
#entities
<!ENTITY % b "xxxxx" "" "" >

#data
<!DOCTYPE a[
<!ENTITY % b "yyy" >
]><a/>
#errors
#document
| <!DOCTYPE a>
| <{} a>
#entities
<!ENTITY % b "yyy" "" "" >

#data
<!DOCTYPE a[
<!ENTITY % b 'yyy' >
]><a/>
#errors
#document
| <!DOCTYPE a>
| <{} a>
#entities
<!ENTITY % b "yyy" "" "" >

#data
<!DOCTYPE a[
<!ENTITY % b "a&#x24;b">
]><a/>
#errors
#document
| <!DOCTYPE a>
| <{} a>
#entities
<!ENTITY % b "a$b" "" "" >

#data
<!DOCTYPE a[
<!ENTITY % b '&#32;'>
]><a/>
#errors
#document
| <!DOCTYPE a>
| <{} a>
#entities
<!ENTITY % b " " "" "" >

#data
<!DOCTYPE a[
<!ENTITY % b "&amp;abc">
]><a/>
#errors
#document
| <!DOCTYPE a>
| <{} a>
#entities
<!ENTITY % b "&amp;abc" "" "" >

#data
<!DOCTYPE a[
<!ENTITY % b "c">
<!ENTITY % b "d">
]><a/>
#errors
3;1;duplicate para entity decl;;b;w
#document
| <!DOCTYPE a>
| <{} a>
#entities
<!ENTITY % b "c" "" "" >

#data
<!DOCTYPE a[
<!ENTITY % b "c">
<!ENTITY b "d">
]><a/>
#errors
#document
| <!DOCTYPE a>
| <{} a>
#entities
<!ENTITY b "d" "" "" >
<!ENTITY % b "c" "" "" >

#data
<!DOCTYPE a[
<!ENTITY % b SYSTEM "x" >
%c;
<!ENTITY % d SYSTEM "y" >
]><a/>
#errors
3;1;stop processing;;;i
3;2;string in internal subset;;;m
#document
| <!DOCTYPE a>
| <{} a>
#entities
| <!ENTITY % b "" "" "x" >

#data
<?xml version="1.0" standalone="yes"?>
<!DOCTYPE a[
<!ENTITY % b SYSTEM "x" >
%c;
<!ENTITY % d SYSTEM "y" >
]><a/>
#errors
4;2;string in internal subset;;;m
#document
| <!DOCTYPE a>
| <{} a>
#entities
| <!ENTITY % b "" "" "x" >
| <!ENTITY % d "" "" "y" >
